Intercept 'end' method
As of Node v16, the 'finish' event of a transform stream is not emitted
when .end() is called. In fact, it appears that is was never safe to
assume it would be. From the docs:

    The 'finish' event is emitted after stream.end() is called and all
    chunks have been processed by stream._transform().

This commit resolves the issue by intercepting the call to res.end()
instead of waiting for a 'finish' event. The callback supplied to the
MockServerResponse constructer is no longer shorthand for adding a
'finish' listener.

close #14